+++ src/components/GIcon/GIcon.vue
@@ -1,169 +1,105 @@
-<template>
-	<i v-if="activeTags.i" :class="iconClass" :style="iconStyle"
-		 :aria-hidden="attributes.ariaHidden"
-		 :role="attributes.role" @click="onClick">{{materialIconName}}</i>
-	<svg v-else-if="activeTags.svg" :class="iconClass" :style="iconStyle"
-			 :aria-hidden="attributes.ariaHidden"
-			 :role="attributes.role"
-			 xmlns="http://www.w3.org/2000/svg"
-			 viewBox="0 0 24 24" @click="onClick">
-		<path :d="icon"></path>
-	</svg>
-	<img v-else-if="activeTags.img" :src="icon" alt="Can't load icon" :class="iconClass" :style="iconStyle" @click="onClick"/>
-	<span v-else-if="activeTags.slot" @click="onClick">
-		<slot></slot>
-	</span>
-</template>
-
 <script>
-  import { computed, ref, onMounted, onUpdated } from '@vue/composition-api';
-  import { convertToUnit } from '../../utils/helpers';
-	import { setBackgroundColor, setTextColor } from '../../mixins/colorable';
-  import { Fragment } from 'vue-fragment'
+  import {computed} from '@vue/composition-api';
+  import {convertToUnit} from '../../utils/helpers';
+  import {setTextColor} from '../../mixins/colorable';
 
   export default {
-    name: 'GIcon',
-    components: { Fragment },
+    name: "GIconJSX",
     props: {
       value: String,
       dense: Boolean,
       disabled: Boolean,
       left: Boolean,
       right: Boolean,
+      color: [String],
       size: [Number, String],
       xSmall: Boolean,
       small: Boolean,
       medium: Boolean,
       large: Boolean,
       xLarge: Boolean,
-      color: [String],
-      tag: {
-        type: String,
-        required: false,
-        default: 'i',
-      },
     },
-
     setup(props, context) {
-      const icon = ref('')
-
-      function getIcon() {
-        icon.value = context.slots.default ? context.slots.default()[0].text.trim() : ''
-      }
-
-      onMounted(() => getIcon())
-      onUpdated(() => getIcon())
-
-      function renderFontAwesomeIcon(nodeData) {
-				nodeData.class[icon.value] = true
-				nodeData.style['fontSize'] = getSize(props)
-			}
-
-      function renderMaterialIcon(nodeData) {
-        let iconType = 'material-icons'
-        const delimiterIndex = icon.value.indexOf('-')
-        const isNotMdiIcon = delimiterIndex <= -1
-
-        if (isNotMdiIcon) {
-          nodeData.materialIcon = icon.value
-        } else {
-          iconType = icon.value.slice(0, delimiterIndex)
-        }
-
-        nodeData.class[iconType] = true
-        nodeData.class[icon.value] = !isNotMdiIcon
-        nodeData.style['fontSize'] = getSize(props)
-      }
-
-      function renderSvgIcon(nodeData) {
-        nodeData.tag.svg = true
-        nodeData.tag.i = false
-
-        nodeData.class['g-icon__svg'] = true
-				nodeData.style['fontSize'] = getSize(props)
-				nodeData.style['width'] = getSize(props)
-				nodeData.style['height'] = getSize(props)
-      }
-
-      function renderCustomSvgIcon(nodeData) {
-        nodeData.tag.img = true
-        nodeData.tag.i = false
-        nodeData.style['fontSize'] = getSize(props) * 2
-        nodeData.style['width'] = getSize(props) * 2
-        nodeData.style['height'] = getSize(props) * 2
-      }
-
-      let nodeData = computed(() => {
-        let _nodeData = {
-          tag: {
-            i: true,
-            svg: false,
-            img: false,
-            slot: true
-          },
-          class: {},
-          style: {},
-          materialIcon: ''
-        }
-
-        if (isFontAwesome5(icon.value)) {
-          renderFontAwesomeIcon(_nodeData)
-        } else if (isSvgPath(icon.value)) {
-          renderSvgIcon(_nodeData)
-        } else if (isCustomSvgIcon(icon.value)) {
-          renderCustomSvgIcon(_nodeData)
-        } else {
-          renderMaterialIcon(_nodeData)
-        }
-        _nodeData.tag.slot = false
-        return _nodeData
-      })
-
-      const materialIconName = computed(() => {
-        return nodeData.value.materialIcon
-      })
-
-      let activeTags = computed(() => ({
-        ...nodeData.value.tag
-      }))
-
-      let iconColor = computed(() => setTextColor(props.color, {}))
-
-      let iconClass = computed(() => ({
-        ...nodeData.value.class,
-        ...iconColor.value.class,
-        'g-icon': true,
-        'g-icon__dense': props.dense,
-        'g-icon__disabled': props.disabled,
-        'g-icon__left': props.left,
-        'g-icon__right': props.right,
-        'g-icon__link': !!context.listeners.click
-      }))
-
-      let iconStyle = computed(() => ({
-        ...nodeData.value.style,
-        ...iconColor.value.style
-      }))
-
-      // accessibility attrs
-      let attributes = computed(() => ({
-        ariaHidden: !!context.listeners.click,
-        role: context.listeners.click ? 'button' : null
-      }))
-
-      let onClick = (event) => {
+      const onClick = (event) => {
         context.emit('click', event);
       }
 
-      return {
-        iconClass,
-        iconStyle,
-        attributes,
-        activeTags,
-        icon,
-        materialIconName,
-        onClick,
+      const iconColor = computed(() => setTextColor(props.color, {}))
+
+      function genFontAwesomeIcon(icon, iconClass, iconStyle) {
+        iconClass.value[icon.value] = true
+        return <i class={iconClass.value}
+                  style={iconStyle.value}
+                  vOn:click={onClick}/>
       }
+
+      function genMaterialIcon(icon, iconClass, iconStyle) {
+        let iconType = 'material-icons'
+        const delimiterIndex = icon.value.indexOf('-')
+        const isMdiIcon = !(delimiterIndex <= -1)
+        if (isMdiIcon) {
+          iconType = icon.value.slice(0, delimiterIndex)
+          iconClass.value[icon.value] = true
+        }
+
+        iconClass.value[iconType] = true
+        return <i class={iconClass.value}
+                  style={iconStyle.value}
+                  vOn:click={onClick}>{!isMdiIcon ? icon.value : ''}</i>
+      }
+
+      function genSvgIcon(icon, iconClass, iconStyle) {
+        iconClass.value['g-icon__svg'] = true
+        iconStyle.value['width'] = getSize(props)
+        iconStyle.value['height'] = getSize(props)
+        return <svg class={iconClass.value}
+                    style={iconStyle.value}
+                    xmlns="http://www.w3.org/2000/svg"
+                    viewBox="0 0 24 24"
+                    vOn:click={onClick}>
+          <path d={icon.value} fill={props.color}/>
+        </svg>
+      }
+
+      function genCustomSvgIcon(icon, iconClass, iconStyle) {
+        iconStyle.value['width'] = getSize(props)
+        iconStyle.value['height'] = getSize(props)
+        return <img class={iconClass.value}
+                    style={iconStyle.value}
+                    src={icon.value}
+                    alt="Can't load icon"
+                    vOn:click={onClick}/>
+      }
+
+      function genIcon() {
+        const iconClass = computed(() => ({
+          ...iconColor.value.class,
+          'g-icon': true,
+          'g-icon__dense': props.dense,
+          'g-icon__disabled': props.disabled,
+          'g-icon__left': props.left,
+          'g-icon__right': props.right,
+          'g-icon__link': !!context.listeners.click,
+        }))
+
+        const iconStyle = computed(() => ({
+          ...iconColor.value.style,
+          'font-size': getSize(props),
+        }))
+        let icon = computed(() => context.slots.default ? context.slots.default() ? context.slots.default()[0].text.trim() : '' : '')
+        return isFontAwesome5(icon.value) ? genFontAwesomeIcon(icon, iconClass, iconStyle)
+            : isSvgPath(icon.value) ? genSvgIcon(icon, iconClass, iconStyle)
+                : isCustomSvgIcon(icon.value) ? genCustomSvgIcon(icon, iconClass, iconStyle)
+                    : genMaterialIcon(icon, iconClass, iconStyle)
+      }
+
+      return {
+        genIcon
+      }
+    },
+
+    render() {
+      return this.genIcon()
     }
   }
 
@@ -191,5 +127,5 @@
 </script>
 
 <style scoped lang="scss">
-	@import "GIcon";
+  @import "GIcon";
 </style>
